job_search_ai/requirements.txt
google-generativeai==0.8.3
customtkinter==5.2.2
Pillow==10.4.0
requests==2.31.0
python-dotenv==1.0.1